+++ companion/src/CMakeLists.txt
@@ -106,7 +106,6 @@ include_directories(
   ${CMAKE_CURRENT_BINARY_DIR}
   ${CMAKE_SOURCE_DIR}
   ${COMPANION_SRC_DIRECTORY}
-  ${COMPANION_SRC_DIRECTORY}/datamodels
   ${COMPANION_SRC_DIRECTORY}/shared
   ${SIMU_SRC_DIRECTORY}
   ${COMPANION_SRC_DIRECTORY}/thirdparty/qcustomplot
@@ -205,6 +204,7 @@ set(CPN_COMMON_LIB common)
 
 ############# Supporting libraries ###############
 
+add_subdirectory(datamodels)
 add_subdirectory(firmwares)
 add_subdirectory(shared)
 add_subdirectory(modeledit)
@@ -249,8 +249,7 @@ set(companion_SRCS
   wizarddata.cpp
   wizarddialog.cpp
   styleeditdialog.cpp
-  datamodels/rawsourceitemmodel.cpp  # TODO move to own lib with other item models
-  dialogs/filesyncdialog.cpp         # also move all dialogs into subfolder
+  dialogs/filesyncdialog.cpp         # TODO move to own lib with other dialogs
   )
 
 set(companion_MOC_HDRS
@@ -284,7 +283,6 @@ set(companion_MOC_HDRS
   modelslist.h
   styleeditdialog.h
   helpers_html.h
-  datamodels/rawsourceitemmodel.h
   dialogs/filesyncdialog.h
   )
