+++ companion/src/modeledit/CMakeLists.txt
@@ -16,7 +16,6 @@ set(modeledit_NAMES
 )
 
 set(modeledit_SRCS
-  switchitemmodel.cpp
   flightmodes.cpp
   inputs.cpp
   mixes.cpp
@@ -29,7 +28,6 @@ set(modeledit_SRCS
 )
 
 set(modeledit_HDRS
-  switchitemmodel.h
   flightmodes.h
   inputs.h
   mixes.h
@@ -59,4 +57,4 @@ qt5_wrap_ui(modeledit_SRCS ${modeledit_UIS})
 qt5_wrap_cpp(modeledit_SRCS ${modeledit_HDRS})
 
 add_library(modeledit ${modeledit_SRCS})
-target_link_libraries(modeledit PRIVATE ${CPN_COMMON_LIB} Qt5::Multimedia)
+target_link_libraries(modeledit PRIVATE datamodels ${CPN_COMMON_LIB} Qt5::Multimedia)
